+++ Software/MATLAB/test/unit/TestStopQuery.m
@@ -20,7 +20,7 @@ classdef TestStopQuery < matlab.unittest.TestCase
     end
     
     methods (TestMethodTeardown)
-        function testTearDown(testCase)
+        function testTearDown(testCase) %#ok<MANU>
             
         end
     end
@@ -49,7 +49,7 @@ classdef TestStopQuery < matlab.unittest.TestCase
 
             queryId = athena.submitQuery(queryStr, testCase.resultBucket);
 
-            queryStatus = athena.getStatusOfQuery(queryId);
+            queryStatus = athena.getStatusOfQuery(queryId); %#ok<NASGU>
             
             [stopStatus, stopMsg] = athena.stopQueryExecution(queryId);
             queryStatus2 = athena.getStatusOfQuery(queryId);
@@ -57,7 +57,8 @@ classdef TestStopQuery < matlab.unittest.TestCase
             testCase.verifyEqual(char(queryStatus2.toString), 'CANCELLED');
             testCase.verifyEqual(stopStatus, true);
             testCase.verifyClass(stopMsg, 'java.util.Optional');
-            testCase.verifyEqual(char(stopMsg.toString), 'Optional[OK]');
+            
+            testCase.verifyTrue(strcmp(char(stopMsg.toString), 'Optional[OK]') || strcmp(char(stopMsg.toString), 'Optional[]'));
             athena.shutdown();
             
         end
@@ -66,10 +67,21 @@ classdef TestStopQuery < matlab.unittest.TestCase
     % Helper methods
     methods
         function doAlternativeInitialize(testCase, ath)
-            reg = getenv('AWS_DEFAULT_REGION')
-            awsKeyId = getenv('AWS_ACCESS_KEY_ID')
-            awsSecretKey = getenv('AWS_SECRET_ACCESS_KEY')
+            % Configure values before running local test e.g.:
+            % setenv('AWS_DEFAULT_REGION', 'us-west-2');
+            % setenv('AWS_ACCESS_KEY_ID', 'AS<REDACTED>7');
+            % setenv('AWS_SECRET_ACCESS_KEY', 'YB<REDACTED>h');
             
+            reg = getenv('AWS_DEFAULT_REGION');
+            testCase.assertNotEmpty(reg, 'There must be a AWS_DEFAULT_REGION environment varaible present for local testing');
+            if ~strcmpi(reg, 'us-west-2')
+                warning('Default unit test data held in: us-west-2, use this region');
+            end
+            awsKeyId = getenv('AWS_ACCESS_KEY_ID');
+            testCase.assertNotEmpty(awsKeyId, 'There must be a AWS_ACCESS_KEY_ID environment varaible present for local testing');
+            awsSecretKey = getenv('AWS_SECRET_ACCESS_KEY');
+            testCase.assertNotEmpty(awsSecretKey, 'There must be a AWS_SECRET_ACCESS_KEY environment varaible present for local testing');
+             
             cp = aws.auth.CredentialProvider.getBasicCredentialProvider(awsKeyId, awsSecretKey);
             cls = 'software.amazon.awssdk.auth.credentials.StaticCredentialsProvider';
             testCase.assertClass(cp, cls, sprintf('The CredentialsProvider should be of type "%s"\n', cls));
@@ -85,7 +97,7 @@ classdef TestStopQuery < matlab.unittest.TestCase
     
 end
 
-function [queryId, S] = syncSubmitQuery(ath, queryStr, resultBucket)
+function [queryId, S] = syncSubmitQuery(ath, queryStr, resultBucket) %#ok<DEFNU>
     
     queryId = ath.submitQuery(queryStr, resultBucket);