+++ RELEASENOTES.md
@@ -1,17 +1,27 @@
-#  MATLAB Interface *for AWS Athena* - Release Notes
+# MATLAB Interface *for AWS Athena* - Release Notes
+
+## Release 0.2.5 (13th Dec 2021)
+
+* Updated pom.xml latest AWS 1.12.128 & 2.17.99 BOMs to include log4j updates
 
 ## Release 0.2.4 (1st Oct 2020)
+
 * Added basic proxy support
 * Updated AWS SDK, junit & Maven Shade versions
 * Add stopQueryExecution method
 
 ## Release 0.2.3 (22th Mar 2021)
+
 * Fixed minor issue with naming of bucket
 
 ## Release 0.2.2 (13th Oct 2020)
+
 * Bumped junit version
 
 ## Release 0.2.1 (13th Mar 2020)
+
 * Updates to SECURITY.md
 
 ## Release 0.2.0 -- Initial public release
+
+[//]: #  (Copyright 2019-2021 The MathWorks, Inc.)